Update build.xml to create a independent PhoneGap project.
config.xml is not longer dependent on the ANT tool.

The ANT tool is now similar to PhoneGap-BlackBerry.
In other words, phonegap-blackberry-widget is not longer home
to a widget project. Instead, it generates standalone projects
that contain the compiled extension and phonegap.js.

Create and Update PhoneGap projects
-----------------------------------

  - ant help explains everything
  - Create a new PhoneGap project
      ant create -Dproject.path="..."
  - Update an existing PhoneGap project
      ant update -Dproject.path="..."

Running a PhoneGap project
--------------------------

  - cd C:\my\project
  - ant help explains everything
  - Run on the simulator
      ant load-simulator
  - Run on the device
      ant load-device
  - A few maintenance tasks
      ant build
      ant clean / clean-simulator / clean-device