Remove reference to FledgeHook in project build file.
The 2.1 update to the BlackBerry WebWorks SDK no longer installs
FledgeHook.exe or FledgeHook.dll.  The original intention of these
files was to manage the lifecycle of a simulator launched from an
Eclipse plugin environment.  It has been confirmed with BlackBerry
WebWorks developers that FledgeHook was intentionally removed from
the latest SDK update.

http://supportforums.blackberry.com/t5/Web-and-WebWorks-Development/FledgeHook-removed-from-latest-SDK/td-p/1189641

Attempting to invoke the load-simulator target in a project will
fail when FledgeHook.exe does not exist.

 - Remove invocation of FledgeHook.exe from project build.xml

 - Change references of Widget to WebWorks to match official name.