Add support for targetWidth and targetHeight to camera API.
Add targetWidth, targetHeight and encodingType options to the camera
API.  If targetWidth or targetHeight is specified or if encodingType
is not the default value of JPEG then post process the image taken
with the camera to match the specified options.

Maintain the pre-existing behavior of the API unless the new options
are specified.  Please note the following concerns if using the new
options.

 - Requires additional processing time before image is returned
   (5+ seconds on a torch).

 - EXIF data from original image is lost so image may no longer
   automatically orient when displayed.

 - The quality option is only honored for JPEG output.

 - Aspect ratio maintained.  Return a new image within the boundaries
   specified by targetWidth / targetHeight but adjust dimensions
   to maintain aspect ratio.