Added debug token support to the debug-device command in qnx
Anyone developing on BlackBerry 10 has surely crossed the annoying issue
where you need to have a unqiue version number everytime you sign an
app.

Added an optional value to project.properties for the device pin (which
you can get from the about screen Hardware section) that when filled in
will create and load a debug token on your device and get around needing
to sign the app to load it on the phone.